requirements.txt
asyncio==3.4.3
asyncpg==0.20.0
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
patsy==0.5.1
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
scikit-posthocs==0.6.2
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
statsmodels==0.10.2
